Moved tasks into their own directoy
